# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API's BASE",
        "description": "API's for transportation management",
        "version": "1.0.0"
    },
    "paths": {
...
        "/base-backend/public/api/person": {
            "get": {
                "tags": [
                    "Person"
                ],
                "summary": "Obtener información con filtros y ordenamiento",
                "operationId": "09150f7d3166d010e067957d5c85470e",
                "parameters": [
                    {
                        "name": "username",
                        "in": "query",
                        "description": "Filtrar por nombre de la persona",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "person.name",
                        "in": "query",
                        "description": "Filtrar por nombre de la persona",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "person.business_name",
                        "in": "query",
                        "description": "Filtrar por nombre de la empresa",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "description": "Fecha de inicio",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "Fecha de fin",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de Empresas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Person"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validación fallida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Person"
                ],
                "summary": "Crear una nueva persona",
                "operationId": "004425aebe301a2b4c028287cbf9145a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "numberDocument"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "persona Ejemplo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "persona creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Person"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Datos inválidos"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/base-backend/public/api/person/{id}": {
            "get": {
                "tags": [
                    "Person"
                ],
                "summary": "Obtener detalles de una persona por ID",
                "operationId": "b702bc4b986e387cf4a4192435d6ab5f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la persona",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "persona encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Person"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Persona no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Persona no encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Person"
                ],
                "summary": "Actualizar la información de una persona",
                "operationId": "f11b99eece87d155b73afce5491d9ee6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la persona a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "persona Ejemplo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "persona actualizada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Person"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Persona no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Persona no encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Datos inválidos"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Person"
                ],
                "summary": "Eliminar persona por ID",
                "operationId": "3914db80ae5eab9d9a973450098be345",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la persona",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "persona eliminada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "persona eliminada exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Persona no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Persona no encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Person": {
                "title": "Person",
                "description": "Person model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "type_document": {
                        "type": "string",
                        "example": "DNI"
                    },
                    "type_person": {
                        "type": "string",
                        "example": "Individual"
                    },
                    "number_document": {
                        "type": "string",
                        "example": "12345678"
                    },
                    "names": {
                        "type": "string",
                        "example": "Juan"
                    },
                    "father_surname": {
                        "type": "string",
                        "example": "Perez"
                    },
                    "mother_surname": {
                        "type": "string",
                        "example": "Lopez"
                    },
                    "business_name": {
                        "type": "string",
                        "example": "Perez & Co."
                    },
                    "address": {
                        "type": "string",
                        "example": "Av. Siempre Viva 123"
                    },
                    "phone": {
                        "type": "string",
                        "example": "987654321"
                    },
                    "email": {
                        "type": "string",
                        "example": "juanperez@example.com"
                    },
                    "occupation": {
                        "type": "string",
                        "example": "Engineer"
                    },
                    "status": {
                        "type": "string",
                        "example": "Activo"
                    },
                    "server_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-02T12:00:00Z"
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": null
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}